refactor: useSnackbar 훅 분리
