+++ src/pages/ItemList.tsx
@@ -1,35 +1,22 @@
-import { useEffect, useCallback, useRef } from 'react';
 import { getItemList } from 'redux/action-creators/itemListThunk';
 import ItemContainer from 'components/ItemList/ItemContainer';
 import styled from 'styled-components';
 import { ItemListAction } from 'redux/actions/itemList';
 import { useAppDispatch } from 'hooks/useAppDispatch';
 import useCartList from 'hooks/useCartList';
-import { useDispatch } from 'react-redux';
-import { SnackbarActionType } from 'redux/actions/snackbar';
 import { useAppSelector } from 'hooks/useAppSelector';
+import useSnackBar from 'hooks/useSnackBar';
+import { useEffect } from 'react';
 
 const ItemList = () => {
   const { data: itemList, error, loading } = useAppSelector(state => state.itemListReducer);
   const dispatch = useAppDispatch<ItemListAction>();
-  // @TODO: 에러 처리
+
   const { cartList, updateCartItemQuantity } = useCartList();
-  const timerRef = useRef(null);
-  const snackbarDispatch = useDispatch();
+  const { openSnackbar } = useSnackBar();
 
   useEffect(() => {
     dispatch(getItemList());
-  }, []);
-
-  const openSnackbar = useCallback(() => {
-    if (timerRef.current) {
-      clearTimeout(timerRef.current);
-      timerRef.current = null;
-    }
-    snackbarDispatch({ type: SnackbarActionType.OPEN_SNACKBAR, payload: 'cart' });
-    timerRef.current = setTimeout(() => {
-      snackbarDispatch({ type: SnackbarActionType.CLOSE_SNACKBAR });
-    }, 3000);
   }, []);
 
   return (
